+++ src/components/button/button_empty/button_empty.test.js
@@ -32,6 +32,15 @@ describe('EuiButtonEmpty', () => {
         expect(component)
           .toMatchSnapshot();
       });
+
+      it('renders a button even when href is defined', () => {
+        const component = render(
+          <EuiButtonEmpty href="#" isDisabled />
+        );
+
+        expect(component)
+          .toMatchSnapshot();
+      });
     });
 
     describe('iconType', () => {